requirements.txt
streamlit==1.32.0
pandas==2.2.0
sqlalchemy==2.0.27
plotly==5.18.0
python-dotenv==1.0.1
openai==1.12.0 